Got rid of using jnilib in favor of standard dylib
Since Mac OS X 10.5, dylib is valid and is now also the default used and
reported by Java. This was only necessary in very early Mac OS X
versions.

Right now it causes problems as some dependent libraries are built with
dylib as expected and we keep insisting on using deprecated pattern.